feat(@webext-pegasus/transport): Added "definePegasusBrowserAPI" API
