+++ packages/transport/devtools.ts
@@ -26,6 +26,7 @@ export function initPegasusTransport(): void {
   );
 
   initTransportAPI({
+    browser: browser,
     emitBroadcastEvent: eventRuntime.emitBroadcastEvent,
     onBroadcastEvent: eventRuntime.onBroadcastEvent,
     onMessage: messageRuntime.onMessage,
