+++ packages/transport/src/BroadcastEventRuntime.ts
@@ -1,14 +1,13 @@
 import type {
   PegasusMessage,
   RuntimeContext,
-  TransportBroadcastEventAPI,
 } from './types';
 import type {JsonValue} from 'type-fest';
 
 import {serializeError} from 'serialize-error';
 import uuid from 'tiny-uid';
 
-import {InternalBroadcastEvent} from './types-internal';
+import {InternalBroadcastEvent, TransportBroadcastEventAPI} from './types-internal';
 
 export interface BroadcastEventRuntime extends TransportBroadcastEventAPI {
   /**
