+++ packages/transport/src/MessageRuntime.ts
@@ -1,9 +1,8 @@
 import type {
   OnMessageCallback,
   RuntimeContext,
-  TransportMessagingAPI,
 } from './types';
-import type {InternalMessage} from './types-internal';
+import type {InternalMessage, TransportMessagingAPI} from './types-internal';
 import type {JsonValue} from 'type-fest';
 
 import {serializeError} from 'serialize-error';
