+++ packages/transport/src/definePegasusBrowserAPI.ts
@@ -1,0 +1,9 @@
+import { type Browser } from 'webextension-polyfill';
+
+import { getTransportAPI } from './TransportAPI';
+
+export function definePegasusBrowserAPI(): Browser | null {
+  const {browser} = getTransportAPI();
+
+  return browser;
+}
